refactored searchBar code
